unicorn_contracts code styling
