+++ .prettierrc.js
@@ -1,6 +1,6 @@
 module.exports = {
   trailingComma: "es5",
-  tabWidth: 2,
+  tabWidth: 4,
   semi: true,
   singleQuote: true,
 };
